feat: ui component 파일 분리(#26)
